pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=1.12.2",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
